backend/requirements.txt
Django==4.2            # Framework web principal (version stable)
djangorestframework==3.15  # API REST avec Django
psycopg2-binary==2.9   # Driver PostgreSQL pour PostGIS
django-cors-headers==4.7  # Gestion du CORS entre React et Django
gunicorn==22.0        # Serveur WSGI pour exécuter Django en production
python-dotenv==1.0 #Read key-value pairs from a .env file and set them as environment variables
dj-database-url==2.3
asgiref==3.6.0       
dj-rest-auth==2.1.9
django-allauth==0.54.0
django-extensions==3.2
drf-yasg==1.21
Pillow==11.0
PyJWT==2.6.0 